+++ todo/src/actions/index.js
@@ -1,0 +1,7 @@
+let nextTodoId = 0
+
+export const addTodo = text => ({
+    type: 'ADD_TODO',
+    id: nextTodoId++,
+    text
+})
