+++ todo/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import App from './App';
+import App from './components/App';
 
 import reducer from './reducer';
 import './index.css';
@@ -16,7 +16,7 @@ const store = createStore(
     window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 );
 
-//First, create store, wrap around App (take state off React, put in Redux)
+//FIRST, create store, wrap around App (take state off React, put in Redux) - pass 'reducer' into createStore(reducer)
 ReactDOM.render(
     <Provider store={store} >
         <App />
